requirements.txt
Flask
flask_cors
gunicorn
python-dotenv
firebase-admin
psycopg2-binary
Flask-SQLAlchemy
pyjwt
bcrypt
boto3
Pillow